feat(game): add activeTurn

+++ src/game.c
@@ -4,7 +4,6 @@
  */
 
 #include "game.h"
-#include "fen.h"
 
 void initGame(Game *game, Fen *fen) {
   game->state = 0;                // Reset state
@@ -15,6 +14,6 @@ void initGame(Game *game, Fen *fen) {
   fenPopulateBoard(fen, game->board);
 }
 
-uint8_t activeTurn(Game *game) {
+int activeTurn(Game *game) {
   return ((game->state & (1 << ACTIVE_COLOR_BIT)) == 0) ? 0 : 1;
 }
